negative add fund fixed

+++ transaction/views.py
@@ -1,3 +1,4 @@
+from django.db.models.fields import DecimalField
 from cards.models import Card
 from django.db import models
 from rest_framework.response import Response
@@ -8,7 +9,7 @@ from rest_framework.permissions import IsAuthenticated
 from .models import Transaction
 from .serializers import TransactionSerializer, AddFundSerializer
 from django.http import Http404
-
+from decimal import Decimal
 # Create your views here.
 class TransactionView(APIView):
     """
@@ -70,7 +71,7 @@ class AddFundView(APIView):
 
             
             serializer.save(destination=request.user.phone, type='Card')
-            request.user.balance = request.user.balance+int(request.data['amount'])
+            request.user.balance = request.user.balance+Decimal(request.data['amount'])
             request.user.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
